% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate_factor_plot.R
\name{automate_factor_plot}
\alias{automate_factor_plot}
\title{Automate the factor_plot workflow when we have many categorical variables}
\usage{
automate_factor_plot(df, yVar, facVars = NULL, ylim = NULL)
}
\arguments{
\item{df}{The data frame}

\item{yVar}{The name of the target variable}

\item{ylim}{}
}
\description{
A interactive interface to display the plots one by one. Meanwhile we can write codes in the
text editor to modify the categorical variables
}
\author{
Jiacheng He
}
