% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate_factor_to_lmfit.R
\name{automate_factor_to_lmfit}
\alias{automate_factor_to_lmfit}
\title{Automate the factor_to_lmfit workflow for all categorical variables in the data frame}
\usage{
automate_factor_to_lmfit(df, yVar, keep = FALSE)
}
\description{
Automatically transform all categorical variables into linear fit
}
\author{
Jiacheng He
}
