% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate_numeric_plot.R
\name{automate_numeric_plot}
\alias{automate_numeric_plot}
\title{Automate the Numeric_plot workflow when we have many numeric variables}
\usage{
automate_numeric_plot(df, yVar, ylim = NULL)
}
\author{
Jiacheng He
}
