% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_linear_fit.R
\name{categorical_linear_fit}
\alias{categorical_linear_fit}
\title{Insert linear projection of categorical variables}
\usage{
categorical_linear_fit(var_list, y, df, replaced = FALSE)
}
\arguments{
\item{var_list}{A list of names of categorical variables (characters)}

\item{y}{The name of the target variable (character)}

\item{df}{The data frame which contains the variables}

\item{replaced}{Whether to replace the original factor cloumns with the fitted value cloumns}
}
\value{
A data frame with new columns of the fitted values.
}
\description{
This function runs a set of linear regressions of y on each of the specified categorical variables. Then insert the fitted
values of the linear regressions
}
\author{
Jiacheng He
}
