% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_to_lmfit.R
\name{categorical_to_lmfit}
\alias{categorical_to_lmfit}
\title{Insert linear projection of categorical variables}
\usage{
categorical_to_lmfit(varName, y_train, df, alpha = 0, lambda = 0)
}
\arguments{
\item{varName}{A character, the name of the categorical variable}

\item{y_train}{The target variable in the training set}

\item{df}{The data frame which contains the variables}

\item{alpha}{alpha=0 for ridge, alpha=1 for LASSO.}

\item{lambda}{Regularization parameter}
}
\value{
A data frame with new columns of the fitted values.
}
\description{
This function runs a set of linear regressions of y on each of the specified categorical variables. Then insert the fitted
values of the linear regressions
}
\author{
Jiacheng He
}
