% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_to_lmfit.R
\name{dummy_to_lmfit}
\alias{dummy_to_lmfit}
\title{Transform dummy variables to linear model fit}
\usage{
dummy_to_lmfit(df, yVar, dummy, trainIndex = NULL, alpha = 0, lambda = 0,
  trCon = NULL, cv_method = "none")
}
\description{
This function is a backend for factor_to_lmfit and text_to_features
}
\author{
Jiacheng He
}
