% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_impute.R
\name{random_impute}
\alias{random_impute}
\title{Randomly impute missing values}
\usage{
random_impute(df, seed = 1000)
}
\description{
Draw from non-missing values, then replace missing values
}
\author{
Jiacheng He
}
