% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_submission.R
\name{save_submission}
\alias{save_submission}
\title{Save the Kaggle submissions .csv files}
\usage{
save_submission(df_sub, path = "submission/")
}
\description{
Automate the numbering and saving of the "submission_.csv" files
}
\author{
Jiacheng He
}
