% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_dtm.R
\name{text_to_dtm}
\alias{text_to_dtm}
\title{Transform a text column from a data frame into document-term matrix}
\usage{
text_to_dtm(df, id, textVar, form, n_gram = 1, stop_rm = TRUE,
  stem = NULL, dtm_value = "tf_idf")
}
\arguments{
\item{df}{The data frame}

\item{id}{The colnumn name of the id variable}

\item{textVar}{The column name of the text variable}
}
\description{
Can perform stemming and removing stop words
}
\author{
Jiacheng He
}
