% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_features.R
\name{text_to_features}
\alias{text_to_features}
\title{Transform text column into numeric features}
\usage{
text_to_features(df, yVar, id, textVar, trainIndex = NULL, n_gram = 1,
  stop_rm = TRUE, stem = NULL, dtm_value = "tf_idf")
}
\arguments{
\item{df}{The data frame}

\item{id}{The colnumn name of the id variable}

\item{col_text}{The column name of the text variable}
}
\description{
Features such as text length, sentiment, and linear model fit
}
\author{
Jiacheng He
}
