% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_lmfit.R
\name{text_to_lmfit}
\alias{text_to_lmfit}
\title{Turn text variable into linear fitted value}
\usage{
text_to_lmfit(df, yVar, id, textVar, form, n_gram = 1, stop_rm = TRUE,
  stem = NULL, dtm_value = "tf_idf", trainIndex = NULL, alpha = 0,
  lambda = 0, trCon = NULL, cv_method = "none", drop = TRUE)
}
\author{
Jiacheng He
}
