% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Save the well-tuned models}
\usage{
write_model(model, model_name, note = NA, models_file = "models.RData",
  log_file = "log.RData", path = "model/")
}
\description{
Gather the models into a list and save them as .RData
}
\author{
Jiacheng He
}
