% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_models_log.R
\name{write_models_log}
\alias{write_models_log}
\title{Save the well-tuned models}
\usage{
write_models_log(model, model_name, file_name = "models_log.RData",
  path = "models/")
}
\description{
Gather the models into a list and save them as .RData
}
\author{
Jiacheng He
}
