% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_submission.R
\name{write_submission}
\alias{write_submission}
\title{Save the Kaggle submissions .csv files}
\usage{
write_submission(df_sub, path = "submission/")
}
\description{
Automate the numbering and saving of the "submission_.csv" files
}
\author{
Jiacheng He
}
