% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blending.R
\name{blending}
\alias{blending}
\title{Blend the tuned models altogether}
\usage{
blending(method, model_list, df_train, df_test, ...)
}
\description{
After training many models, use a final model to combine them all together
}
\author{
Jiacheng He
}
