% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_plot.R
\name{factor_plot}
\alias{factor_plot}
\title{Plot the counts of each category or the average values of y for each category}
\usage{
factor_plot(df, facVar, yVar = NULL, stat = "identity")
}
\author{
Jiacheng He
}
